recent_conversation: Move click handlers out to an initialize function.
We are moving in the direction of removing click_handlers library.
